requirements.txt
pyyaml
dash
dash_bootstrap_components
plotly
pandas
numpy
openpyxl
pymongo
pymongo[srv]
devtools
pydantic
pydantic_cli
pydantic[email]
phonenumbers
email-validator
dash_ag_grid
python-docx
gunicorn
rich
bcrypt
fastapi
uvicorn
python-jose[cryptography]
python-multipart